Move displaying games to game utils
